chore: 배경 dimmed 컴포넌트로 분리
